requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
slackclient==2.9.4
yarl==1.9.4
